add stats to country page
